requirements_optimized.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
prometheus-client==0.19.0
structlog==23.2.0
psutil==5.9.6
blinker==1.6.3
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
PyMySQL==1.1.0
SQLAlchemy-Utils==0.41.1
python-dateutil==2.8.2
six==1.16.0
typing-extensions==4.8.0
watchdog==3.0.0
websockets==12.0